+++ collective/teaser/browser/common.py
@@ -65,8 +65,8 @@ def get_teasers(context, data, request):
     teaser_list = []
     for teaser in choosen_teasers:
         img_text_part = not show_desc and teaser.Description() or ''
-        img_text = '%s%s' % (teaser.title,
-                         img_text_part and ' - %s' % img_text_part or '')
+        img_text = '%s%s' % (teaser.title.encode('utf-8'),
+                             img_text_part and ' - %s' % img_text_part or '')
         teaser_list.append({
             'title': show_title and teaser.title or None,
             'image': getattr(teaser, 'image', False) \